+++ sem2/lab2/task/input.h
@@ -3,6 +3,6 @@
 #include "../structs/task.h"
 
 int isInt(const char *s);
-size_t strToInt(const char *s);
+long strToInt(const char *s);
 Queue * input_array(size_t size);
 #endif //SEM1_LAB4_INPUT_H
